Takeaway-test————用户修改密码(以及对之前根据id查询员工的字段类型进行修改)

+++ server/src/main/java/com/sky/mapper/EmployeeMapper.java
@@ -50,6 +50,6 @@ public interface EmployeeMapper {
      * @return
      */
     @Select("select * from employee where id = #{id}")
-    Employee QueryById(Integer id);
+    Employee QueryById(Long id);
 
 }
